Working code with java 21+ (Osx Thread.ofVirtual limitation)
