+++ server/netty-server/src/test/scala/sttp/tapir/server/netty/NettyFutureServerBaseTest.scala
@@ -29,7 +29,7 @@ abstract class NettyFutureServerBaseTest(multipart: Boolean) extends TestSuite w
           val createServerTest = new DefaultCreateServerTest(backend, interpreter)
 
           val tests =
-            new AllServerTests(createServerTest, interpreter, backend, multipart = multipart).tests() ++
+            new AllServerTests(createServerTest, interpreter, backend, multipart = multipart, partOtherHeaderSupport = false).tests() ++
               new ServerGracefulShutdownTests(createServerTest, Sleeper.futureSleeper).tests() ++
               new NettyFutureRequestTimeoutTests(eventLoopGroup, backend).tests() ++
               additionalTests(backend)
@@ -57,7 +57,7 @@ abstract class NettyFutureServerBaseTest(multipart: Boolean) extends TestSuite w
         .map(_.port)
         .use { port =>
           (0 until 10000).toList
-            .traverse_ { i =>
+            .traverse_ { _ =>
               // the query is causing a decode failure, which in turn returns a 400 bad request
               // the request body is never read and should be ignored
               basicRequest.response(asStringAlways).body("abcde").post(uri"http://localhost:$port?x=abc").send(backend).map { response =>
