+++ server/tests/src/main/scala/sttp/tapir/server/tests/AllServerTests.scala
@@ -27,7 +27,8 @@ class AllServerTests[F[_], OPTIONS, ROUTE](
     oneOfBody: Boolean = true,
     cors: Boolean = true,
     options: Boolean = true,
-    maxContentLength: Boolean = true
+    maxContentLength: Boolean = true,
+    partOtherHeaderSupport: Boolean = true
 )(implicit
     m: MonadError[F]
 ) {
@@ -38,7 +39,13 @@ class AllServerTests[F[_], OPTIONS, ROUTE](
       (if (file) new ServerFileTests(createServerTest).tests() else Nil) ++
       (if (mapping) new ServerMappingTests(createServerTest).tests() else Nil) ++
       (if (metrics) new ServerMetricsTest(createServerTest, serverInterpreter).tests() else Nil) ++
-      (if (multipart) new ServerMultipartTests(createServerTest, maxContentLengthSupport = maxContentLength).tests() else Nil) ++
+      (if (multipart)
+         new ServerMultipartTests(
+           createServerTest,
+           maxContentLengthSupport = maxContentLength,
+           partOtherHeaderSupport = partOtherHeaderSupport
+         ).tests()
+       else Nil) ++
       (if (oneOf) new ServerOneOfTests(createServerTest).tests() else Nil) ++
       (if (reject) new ServerRejectTests(createServerTest, serverInterpreter).tests() else Nil) ++
       (if (staticContent) new ServerFilesTests(serverInterpreter, backend).tests() else Nil) ++
